requirements.txt
selenium
webdriver-manager
beautifulsoup4
requests-toolbelt
PyYAML